Clase bicicleta y cambios en tipos de datos

+++ Clase2EjercicioGrupal/Aereos/HelicopteroGrua.cs
@@ -9,11 +9,11 @@ namespace Clase2EjercicioGrupal.Aereos
 {
     public class HelicopteroGrua:Aeronaves
     {
-        public HelicopteroGrua(string nombre, string marca, string tamaño, bool usoMilitar, int pesoDeCarga) : base(nombre, marca, tamaño, usoMilitar)
+        public HelicopteroGrua(string nombre, string marca, string tamaño, bool usoMilitar, float pesoDeCarga) : base(nombre, marca, tamaño, usoMilitar)
         {
             PesoDeCarga = pesoDeCarga;
         }
-        public int PesoDeCarga { get; set; }
+        public float PesoDeCarga { get; set; }
 
     }
 }
